+++ model/refresh_token.go
@@ -5,7 +5,7 @@ import (
 )
 
 type RefreshToken struct {
-	ID          float64
+	ID          int64
 	Token       string
 	ExpiresAt   time.Time
 	CreatedAt   time.Time
